So this makes it not crash but it just crashes later instead..

+++ HPL/sources/math/CRC.h
@@ -22,7 +22,7 @@
 namespace hpl {
 	//----------------------------------------
 
-	typedef unsigned long long tCRCKey;
+	typedef unsigned long tCRCKey;
 
 	//----------------------------------------
